added new feat

+++ Web/Models/CartItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class CartItemModel
     {
-        public long ProductID { get; set; }
+        public int ProductID { get; set; }
         public string ProductName { get; set; }
         public int Quantity { get; set; }
         public decimal Price { get; set; }
